extracted newsletter signup form code and made it under the /components directory
